For multi-locale libraries, use the C++ compiler as linker, and...
Adjust the Makefiles for multi-locale libraries to use the C++
compiler as the linker on OSX. For some reason the backend is
throwing the `-shared` flag, which the OSX linker balks at.

Adjust the `.lastcompopts` for muli-locale library C tests to stop
throwing `compileline --libraries`. We don't need to link against
the runtime and third party libraries when building the client
library, so don't bother.

Also remove the last instance of linking against the multi-locale
library server, since this causes the OSX linker to try and
resolve symbols in the Chapel runtime which come from the generated
binary (which is not linked in).

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>