Fix bugs affecting codegen of 'FunctionType' under C backend
After merging #27888, some procedure-pointer tests started failing
under the C backend. This revealed a bunch of problems (some of
which are frankly, resolution-related TODOs).

This PR fixes a laundry-list of things required to get failing tests
to compile under the C backend. Ultimately, it makes procedure-pointer
types behave in a more consistent and uniform manner.

* Refactored pieces of 'isDefAndOrUse' related to (in-)direct calls
* Added new static functions to construct 'Qualifier'
* Adjusted function types to print anonymous formals as '_'
* Added new pass to streamline 'FunctionType' before code generation
* Added new pass to resolve intents for 'FunctionType' formals
* Added way to easily iterate over 'FunctionType' formals
* Added way to run new-style passes over all global symbol vectors

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>