Take first steps towards compiling runtime as dynamic library
Add an environment variable 'CHPL_RT_COMPILE_DYNAMIC_LIBRARY' which
can be used to toggle compiling the runtime as a dynamic library.

Add a new header to the runtime called 'chpl-program-registration.h'.
It contains a struct with fields in it that I collected by trying
to compile the runtime as a dynamic library with strict linker errors.

When you throw '--no-undefined' to the linux linker it will not allow
you to build a shared library where an extern symbol does not have a
definition at link-time. This emulates the strict behavior that is
the default on OSX.

The collected symbols are symbols that the compiled Chapel program is
supposed to provide today. The next step is to replace all the uses of
these symbols with references to the fields of a singleton struct.

When that is done, then the runtime should be able to compile as a
shared library without issues. Note that this is only the first step
towards dynamically loading Chapel programs.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>