Get the 'COMM=gasnet' runtime to link with '-Wl,--no-undefined'
This required locating the source line of all undefined reference
errors that were emitted by the linux linker and replacing those
uses with struct field uses. Then, I looked for the declaration of
the offending symbol in the header files and deleted it.

For QIO, this ended up causing me to remove the entire header file.
I'm of mixed feeling about this, however that API does have a TODO
about moving more QIO code to the Chapel modules, so perhaps it's OK.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>